src/test/java/liveAttack/requirements.txt
requests==2.24.0 
paho-mqtt==1.3.0
CoAPthon3==1.0.1
aio-pika==6.6.1
smbprotocol==1.1.0